src/apps/requirements.txt
# Requirements for src/utils module
# Generated based on analysis of all Python files in src/utils directory

# Core Python libraries (standard library - no version needed)
# asyncio, os, json, pathlib, typing, dataclasses, abc, enum, functools
# threading, tempfile, datetime, hashlib, mimetypes, base64, fnmatch
# inspect, re, sys, time, random, copy, concurrent.futures, io, shutil, uuid

# Third-party packages with versions from main requirements.txt

# HTTP and networking
httpx==0.28.1
requests==2.32.4
aiohttp==3.12.15

# MCP (Model Context Protocol)
mcp==1.9.4

# AI/ML libraries
anthropic==0.55.0
openai==1.90.0
python-dotenv==1.0.1

# Data processing and manipulation
pandas==2.3.0
numpy==2.2.6
scipy==1.15.3

# Image processing
pillow==11.3.0

# File format handling
PyYAML==6.0.2
toml==0.10.2
tomlkit==0.13.3
xmltodict==0.14.2
configobj==5.0.9
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12

# Document processing
mammoth==1.11.0
markdownify==1.2.0
pdfminer==20191125
pdfminer.six==20250506
PyPDF2==3.0.1
python-pptx==1.0.2
beautifulsoup4==4.13.4
lxml==6.0.0

# File system and utilities
fsspec==2025.9.0
unidiff==0.7.5

# Logging and monitoring
loguru==0.7.2
codetiming==1.4.0

# Async and concurrency
tqdm==4.67.1

# Retry mechanisms
retry==0.9.2

# Gymnasium (for RL environments)
gymnasium==1.2.0

# Web scraping and parsing
puremagic==1.30

# Development and testing utilities
pytest==8.4.1
